Moved tests to dedicated directory
