+++ .travis.yml
@@ -1,10 +1,18 @@
+addons:
+    apt:
+        packages:
+            - cloc
 language: python
 python:
     - 2.7
     - 3.3
     - 3.4
     - 3.5
-install: make devenv
+install:
+    - make devenv
+    - export PYTHONPATH=/home/travis/build/standage/aeneas/:$PYTHONPATH
 script:
     - make test
     - make style
+after_success:
+    - make loc
