+++ tests/testdata/grape-cpgat-sorted.gff3
