+++ tests/testdata/lhum-cds-strand.gff3
