+++ tests/testdata/lhum-mrna-span.gff3
