+++ tests/testdata/pbar-withseq.gff3
